Define base keymap setting in welcome crate
