+++ crates/settings/src/settings_file.rs
@@ -1,6 +1,6 @@
 use crate::{
-    settings_store::parse_json_with_comments, settings_store::SettingsStore, KeymapFileContent,
-    Setting, Settings, DEFAULT_SETTINGS_ASSET_PATH,
+    settings_store::parse_json_with_comments, settings_store::SettingsStore, Setting, Settings,
+    DEFAULT_SETTINGS_ASSET_PATH,
 };
 use anyhow::Result;
 use assets::Assets;
@@ -76,43 +76,6 @@ pub fn watch_config_file(
     rx
 }
 
-pub fn handle_keymap_file_changes(
-    mut user_keymap_file_rx: mpsc::UnboundedReceiver<String>,
-    cx: &mut AppContext,
-) {
-    cx.spawn(move |mut cx| async move {
-        let mut settings_subscription = None;
-        while let Some(user_keymap_content) = user_keymap_file_rx.next().await {
-            if let Ok(keymap_content) =
-                parse_json_with_comments::<KeymapFileContent>(&user_keymap_content)
-            {
-                cx.update(|cx| {
-                    cx.clear_bindings();
-                    KeymapFileContent::load_defaults(cx);
-                    keymap_content.clone().add_to_cx(cx).log_err();
-                });
-
-                let mut old_base_keymap = cx.read(|cx| cx.global::<Settings>().base_keymap.clone());
-                drop(settings_subscription);
-                settings_subscription = Some(cx.update(|cx| {
-                    cx.observe_global::<Settings, _>(move |cx| {
-                        let settings = cx.global::<Settings>();
-                        if settings.base_keymap != old_base_keymap {
-                            old_base_keymap = settings.base_keymap.clone();
-
-                            cx.clear_bindings();
-                            KeymapFileContent::load_defaults(cx);
-                            keymap_content.clone().add_to_cx(cx).log_err();
-                        }
-                    })
-                    .detach();
-                }));
-            }
-        }
-    })
-    .detach();
-}
-
 pub fn handle_settings_file_changes(
     mut user_settings_file_rx: mpsc::UnboundedReceiver<String>,
     cx: &mut AppContext,
@@ -183,181 +146,4 @@ pub fn update_settings_file<T: Setting>(
         anyhow::Ok(())
     })
     .detach_and_log_err(cx);
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use fs::FakeFs;
-    use gpui::{actions, elements::*, Action, Entity, TestAppContext, View, ViewContext};
-    use theme::ThemeRegistry;
-
-    struct TestView;
-
-    impl Entity for TestView {
-        type Event = ();
-    }
-
-    impl View for TestView {
-        fn ui_name() -> &'static str {
-            "TestView"
-        }
-
-        fn render(&mut self, _: &mut ViewContext<Self>) -> AnyElement<Self> {
-            Empty::new().into_any()
-        }
-    }
-
-    #[gpui::test]
-    async fn test_base_keymap(cx: &mut gpui::TestAppContext) {
-        let executor = cx.background();
-        let fs = FakeFs::new(executor.clone());
-
-        actions!(test, [A, B]);
-        // From the Atom keymap
-        actions!(workspace, [ActivatePreviousPane]);
-        // From the JetBrains keymap
-        actions!(pane, [ActivatePrevItem]);
-
-        fs.save(
-            "/settings.json".as_ref(),
-            &r#"
-            {
-                "base_keymap": "Atom"
-            }
-            "#
-            .into(),
-            Default::default(),
-        )
-        .await
-        .unwrap();
-
-        fs.save(
-            "/keymap.json".as_ref(),
-            &r#"
-            [
-                {
-                    "bindings": {
-                        "backspace": "test::A"
-                    }
-                }
-            ]
-            "#
-            .into(),
-            Default::default(),
-        )
-        .await
-        .unwrap();
-
-        cx.update(|cx| {
-            let mut store = SettingsStore::default();
-            store.set_default_settings(&test_settings(), cx).unwrap();
-            cx.set_global(store);
-            cx.set_global(ThemeRegistry::new(Assets, cx.font_cache().clone()));
-            cx.add_global_action(|_: &A, _cx| {});
-            cx.add_global_action(|_: &B, _cx| {});
-            cx.add_global_action(|_: &ActivatePreviousPane, _cx| {});
-            cx.add_global_action(|_: &ActivatePrevItem, _cx| {});
-
-            let settings_rx = watch_config_file(
-                executor.clone(),
-                fs.clone(),
-                PathBuf::from("/settings.json"),
-            );
-            let keymap_rx =
-                watch_config_file(executor.clone(), fs.clone(), PathBuf::from("/keymap.json"));
-
-            handle_keymap_file_changes(keymap_rx, cx);
-            handle_settings_file_changes(settings_rx, cx);
-        });
-
-        cx.foreground().run_until_parked();
-
-        let (window_id, _view) = cx.add_window(|_| TestView);
-
-        // Test loading the keymap base at all
-        assert_key_bindings_for(
-            window_id,
-            cx,
-            vec![("backspace", &A), ("k", &ActivatePreviousPane)],
-            line!(),
-        );
-
-        // Test modifying the users keymap, while retaining the base keymap
-        fs.save(
-            "/keymap.json".as_ref(),
-            &r#"
-            [
-                {
-                    "bindings": {
-                        "backspace": "test::B"
-                    }
-                }
-            ]
-            "#
-            .into(),
-            Default::default(),
-        )
-        .await
-        .unwrap();
-
-        cx.foreground().run_until_parked();
-
-        assert_key_bindings_for(
-            window_id,
-            cx,
-            vec![("backspace", &B), ("k", &ActivatePreviousPane)],
-            line!(),
-        );
-
-        // Test modifying the base, while retaining the users keymap
-        fs.save(
-            "/settings.json".as_ref(),
-            &r#"
-            {
-                "base_keymap": "JetBrains"
-            }
-            "#
-            .into(),
-            Default::default(),
-        )
-        .await
-        .unwrap();
-
-        cx.foreground().run_until_parked();
-
-        assert_key_bindings_for(
-            window_id,
-            cx,
-            vec![("backspace", &B), ("[", &ActivatePrevItem)],
-            line!(),
-        );
-    }
-
-    fn assert_key_bindings_for<'a>(
-        window_id: usize,
-        cx: &TestAppContext,
-        actions: Vec<(&'static str, &'a dyn Action)>,
-        line: u32,
-    ) {
-        for (key, action) in actions {
-            // assert that...
-            assert!(
-                cx.available_actions(window_id, 0)
-                    .into_iter()
-                    .any(|(_, bound_action, b)| {
-                        // action names match...
-                        bound_action.name() == action.name()
-                    && bound_action.namespace() == action.namespace()
-                    // and key strokes contain the given key
-                    && b.iter()
-                        .any(|binding| binding.keystrokes().iter().any(|k| k.key == key))
-                    }),
-                "On {} Failed to find {} with key binding {}",
-                line,
-                action.name(),
-                key
-            );
-        }
-    }
 }
